# Change database synchronous mode during fast catchup