--- a/ledger/ledger.go
+++ b/ledger/ledger.go
@@ -32,6 +32,7 @@ import (
 	"github.com/algorand/go-algorand/data/bookkeeping"
 	"github.com/algorand/go-algorand/data/transactions"
 	"github.com/algorand/go-algorand/logging"
+	"github.com/algorand/go-algorand/util/db"
 )
 
 // Ledger is a database storing the contents of the ledger.
@@ -51,6 +52,12 @@ type Ledger struct {
 	// archival determines whether the ledger keeps all blocks forever
 	// (archival mode) or trims older blocks to save space (non-archival).
 	archival bool
+
+	// the synchronous mode that would be used for the ledger databases.
+	synchronousMode db.SynchronousMode
+
+	// the synchronous mode that would be used while the accounts database is being rebuilt.
+	accountsRebuildSynchronousMode db.SynchronousMode
 
 	// genesisHash stores the genesis hash for this ledger.
 	genesisHash crypto.Digest
@@ -89,11 +96,13 @@ func OpenLedger(
 ) (*Ledger, error) {
 	var err error
 	l := &Ledger{
-		log:             log,
-		archival:        cfg.Archival,
-		genesisHash:     genesisInitState.GenesisHash,
-		genesisAccounts: genesisInitState.Accounts,
-		genesisProto:    config.Consensus[genesisInitState.Block.CurrentProtocol],
+		log:                            log,
+		archival:                       cfg.Archival,
+		genesisHash:                    genesisInitState.GenesisHash,
+		genesisAccounts:                genesisInitState.Accounts,
+		genesisProto:                   config.Consensus[genesisInitState.Block.CurrentProtocol],
+		synchronousMode:                db.SynchronousMode(cfg.LedgerSynchronousMode),
+		accountsRebuildSynchronousMode: db.SynchronousMode(cfg.AccountsRebuildSynchronousMode),
 	}
 
 	l.headerCache.maxEntries = 10
@@ -104,7 +113,7 @@ func OpenLedger(
 		}
 	}()
 
-	l.trackerDBs, l.blockDBs, err = openLedgerDB(dbPathPrefix, dbMem)
+	l.trackerDBs, l.blockDBs, err = l.openLedgerDB(dbPathPrefix, dbMem)
 	if err != nil {
 		err = fmt.Errorf("OpenLedger.openLedgerDB %v", err)
 		return nil, err
@@ -113,6 +122,8 @@ func OpenLedger(
 	l.trackerDBs.wdb.SetLogger(log)
 	l.blockDBs.rdb.SetLogger(log)
 	l.blockDBs.wdb.SetLogger(log)
+
+	l.setSynchronousMode(context.Background(), l.synchronousMode)
 
 	err = l.blockDBs.wdb.Atomic(func(ctx context.Context, tx *sql.Tx) error {
 		return initBlocksDB(tx, l, []bookkeeping.Block{genesisInitState.Block}, cfg.Archival)
@@ -207,7 +218,7 @@ func (l *Ledger) verifyMatchingGenesisHash() (err error) {
 	return
 }
 
-func openLedgerDB(dbPathPrefix string, dbMem bool) (trackerDBs dbPair, blockDBs dbPair, err error) {
+func (l *Ledger) openLedgerDB(dbPathPrefix string, dbMem bool) (trackerDBs dbPair, blockDBs dbPair, err error) {
 	// Backwards compatibility: we used to store both blocks and tracker
 	// state in a single SQLite db file.
 	var trackerDBFilename string
@@ -237,7 +248,28 @@ func openLedgerDB(dbPathPrefix string, dbMem bool) (trackerDBs dbPair, blockDBs 
 	if err != nil {
 		return
 	}
+
 	return
+}
+
+// setSynchronousMode sets the writing database connections syncronous mode to the specified mode
+func (l *Ledger) setSynchronousMode(ctx context.Context, synchronousMode db.SynchronousMode) {
+	if synchronousMode < db.SynchronousModeOff || synchronousMode > db.SynchronousModeExtra {
+		l.log.Warnf("ledger.setSynchronousMode unable to set syncronous mode : configured value %d is invalid", l.synchronousMode)
+		return
+	}
+
+	err := l.blockDBs.wdb.SetSynchronousMode(ctx, l.synchronousMode, l.synchronousMode >= db.SynchronousModeFull)
+	if err != nil {
+		l.log.Warnf("ledger.setSynchronousMode unable to set syncronous mode on blocks db: %v", err)
+		return
+	}
+
+	l.trackerDBs.wdb.SetSynchronousMode(ctx, l.synchronousMode, l.synchronousMode >= db.SynchronousModeFull)
+	if err != nil {
+		l.log.Warnf("ledger.setSynchronousMode unable to set syncronous mode on trackers db: %v", err)
+		return
+	}
 }
 
 // initBlocksDB performs DB initialization: