--- a/util/db/fullfsync_darwin.go
+++ b/util/db/fullfsync_darwin.go
@@ -18,5 +18,5 @@ package db
 
 func init() {
 	// See https://www.sqlite.org/pragma.html#pragma_fullfsync
-	initStatements = append(initStatements, "PRAGMA fullfsync=true")
+	enableFullfsyncStatements = append(enableFullfsyncStatements, "PRAGMA fullfsync=true")
 }