## Summary

Reduce the database synchronization for the duration of the fast catchup. Reducing the synchronization to Normal, would make it possible to loose a transaction in case of a power failure. This is not crucial for the fast catchup, since we restart loading the accounts into the database when the algod process restarts.

Note that the database performance gains here are substantial - local tests shown 2x-2.5x performance gains.


## Perf gains

Before:
```
BenchmarkRestoringFromCatchpointFile/Restore-102400-8         	  102400	     64230 ns/op	     72418 ns/last_64k_account
BenchmarkRestoringFromCatchpointFile/Restore-204800-8         	  204800	     71893 ns/op	     89333 ns/last_64k_account
BenchmarkRestoringFromCatchpointFile/Restore-409600-8         	  409600	     82651 ns/op	    106174 ns/last_64k_account
```

After:
```
BenchmarkRestoringFromCatchpointFile/Restore-102400-8         	  102400	     26852 ns/op	     32910 ns/last_64k_account
BenchmarkRestoringFromCatchpointFile/Restore-204800-8         	  204800	     33752 ns/op	     44747 ns/last_64k_account
BenchmarkRestoringFromCatchpointFile/Restore-409600-8         	  409600	     40365 ns/op	     60342 ns/last_64k_account
```

## Test Plan

Added a test unit for the new database functionality